LP-486: Update flutter demo with Douglas Poveda Comments
